requirements.txt
faker
pandas
numpy
seaborn
matplotlib
sqlalchemy
psycopg2
scikit-learn
jupyterlab
tqdm
python-dotenv